refactor: 컴포넌트 폴더 안에 파일명 index.tsx, style.tsx로 통일
