+++ frontend/src/components/common/SelectOptionDescription/style.tsx
