--- frontend/src/components/common/SelectOptionImageBox/SelectOptionImageBox.tsx
@@ -1,14 +1,0 @@
-import * as style from './SelectOptionImageBox.style';
-
-interface SelectOptionImageBoxProps {
-  image: string;
-}
-
-export function SelectOptionImageBox({ image }: SelectOptionImageBoxProps) {
-  return (
-    <>
-      <style.Image src={image} />
-      <style.Button>옵션 위치 보기</style.Button>
-    </>
-  );
-}
