+++ frontend/src/components/common/SelectOptionImageBox/style.tsx
